requirements.txt
fastapi==0.110.0
httpx==0.27.0
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
starlette<0.37.2
uvicorn==0.28.0